// frontend/src/api/generated/models/expenseTrackerDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Backend API
 * A simple ASP.NET Core Web API template with best practices for building scalable and maintainable applications.
 * OpenAPI spec version: v1
 */

export interface ExpenseTrackerDto {
  id: number;
  amount: number;
  category: string;
  /** @nullable */
  description?: string | null;
  date: string;
  userId: number;
  submittedBy: string;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}